[PLAYER-1934] Move BottomOverlay component to its own directory
Component moved in a separate commit to keep Git history.
