+++ sdk/react/src/BottomOverlay/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import BottomOverlay from './BottomOverlay';
+
+export default BottomOverlay;
